started transStringToOperation()
I have created transStringToOperation() function so that when
user presses the = this function will take a string, transform it
into mathematic operation that JS will understand and will output
an answet to the #display > #answer div.
for now I only have coded the function to operate + sign only but
will follow up with rest of the operatros with the same method.
in the future I will have to come up with solution to when user uses
'.' more than once or when more than one operator is pressed/used.